Существующий код разделен на модули. Созданы файлы под будущие модули.
